# swagger.yaml
openapi: 3.0.1
info:
  title: Get Science API
  description: API платформы для научных и образовательных мероприятий
  version: 1.0.0
  contact:
    name: Get Science Team
    url: https://get-science-events.ru
servers:
  - url: https://get-science-events.ru/api/v1
    description: Основной сервер
...
paths:
...
  /applications/{applicationId}:
    put:
      tags:
        - applications
      summary: Обновление существующей заявки
      description: Обновляет существующую заявку
      operationId: updateApplication
      parameters:
        - name: applicationId
          in: path
          description: Идентификатор заявки
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                applicationRequest:
                  $ref: '#/components/schemas/ApplicationRequest'
                fileApplicationFileMetadataDTO:
                  type: array
                  items:
                    $ref: '#/components/schemas/ApplicationFileMetadataDTO'
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: Успешное обновление заявки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '400':
          description: Ошибка валидации данных
        '404':
          description: Заявка не найдена
    get:
      tags:
        - applications
      summary: Получение информации о заявке
      description: Получает информацию о заявке по её идентификатору
      operationId: getApplicationById
      parameters:
        - name: applicationId
          in: path
          description: Идентификатор заявки
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация о заявке успешно получена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '404':
          description: Заявка не найдена
    delete:
      tags:
        - applications
      summary: Удаление заявки
      description: Удаляет заявку по её идентификатору
      operationId: deleteApplication
      parameters:
        - name: applicationId
          in: path
          description: Идентификатор заявки
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Заявка успешно удалена
          content:
            application/json:
              schema:
                type: boolean
        '404':
          description: Заявка не найдена
...
  /files/applications:
    post:
      tags:
        - files
      summary: Создание метаданных файла заявки
      description: Создает метаданные для файла, прикрепленного к заявке
      operationId: createFileApplication
      parameters:
        - name: applicationId
          in: query
          description: Идентификатор заявки
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileApplicationRequest'
      responses:
        '200':
          description: Метаданные успешно созданы
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileApplicationResponse'
        '400':
          description: Ошибка валидации данных
        '404':
          description: Заявка не найдена
  /files/applications/{applicationId}:
    get:
      tags:
        - files
      summary: Получение файлов заявки
      description: Получает список файлов, прикрепленных к заявке
      operationId: getFileApplicationsByApplication
      parameters:
        - name: applicationId
          in: path
          description: Идентификатор заявки
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Список файлов успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileApplicationResponse'
        '404':
          description: Заявка не найдена
  /files/applications/download/{fileId}/direct-link:
    get:
      tags:
        - files
      summary: Получение прямой ссылки на файл заявки
      description: Получает предварительно подписанный URL для скачивания файла заявки
      operationId: getPresignedUrlForApplicationFile
      parameters:
        - name: fileId
          in: path
          description: Идентификатор файла
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: URL успешно получен
          content:
            application/json:
              schema:
                type: string
        '404':
          description: Файл не найден
  /files/applications/download/{fileId}:
    get:
      tags:
        - files
      summary: Скачивание файла заявки
      description: Скачивает файл, прикрепленный к заявке
      operationId: downloadApplicationFile
      parameters:
        - name: fileId
          in: path
          description: Идентификатор файла
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Файл успешно скачан
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Файл не найден
  /files/events/download/{fileId}/direct-link:
    get:
      tags:
        - files
      summary: Получение прямой ссылки на файл мероприятия
      description: Получает предварительно подписанный URL для скачивания файла мероприятия
      operationId: getPresignedUrlForEventFile
      parameters:
        - name: fileId
          in: path
          description: Идентификатор файла
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: URL успешно получен
          content:
            application/json:
              schema:
                type: string
        '404':
          description: Файл не найден
  /files/events/download/{fileId}:
    get:
      tags:
        - files
      summary: Скачивание файла мероприятия
      description: Скачивает файл, прикрепленный к мероприятию
      operationId: downloadEventFile
      parameters:
        - name: fileId
          in: path
          description: Идентификатор файла
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Файл успешно скачан
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Файл не найден
  /files/applications/{fileId}:
    delete:
      tags:
        - files
      summary: Удаление файла заявки
      description: Удаляет файл, прикрепленный к заявке
      operationId: deleteFile
      parameters:
        - name: fileId
          in: path
          description: Идентификатор файла
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Файл успешно удален
          content:
            application/json:
              schema:
                type: boolean
        '404':
          description: Файл не найден
  /files/applications/{applicationId}/doc-required/{docRequiredId}:
    delete:
      tags:
        - files
      summary: Удаление файлов по идентификатору требуемого документа
      description: Удаляет файлы заявки, соответствующие указанному требуемому документу
      operationId: deleteFilesByDocRequiredId
      parameters:
        - name: applicationId
          in: path
          description: Идентификатор заявки
          required: true
          schema:
            type: integer
            format: int64
        - name: docRequiredId
          in: path
          description: Идентификатор требуемого документа
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Файлы успешно удалены
          content:
            application/json:
              schema:
                type: boolean
        '404':
          description: Заявка или требуемый документ не найдены
  /notifications:
    get:
      tags:
        - notifications
      summary: Получение уведомлений
      description: Получает список уведомлений с пагинацией
      operationId: getNotifications
      parameters:
        - name: page
          in: query
          description: Номер страницы (начиная с 0)
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Размер страницы
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Список уведомлений успешно получен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageableResponseNotification'
        '401':
          description: Не авторизован
    delete:
      tags:
        - notifications
      summary: Удаление всех уведомлений
      description: Удаляет все уведомления текущего пользователя
      operationId: deleteAllNotifications
      responses:
        '204':
          description: Все уведомления успешно удалены
        '401':
          description: Не авторизован
  /notifications/unread/count:
    get:
      tags:
        - notifications
      summary: Получение количества непрочитанных уведомлений
      description: Получает количество непрочитанных уведомлений для текущего пользователя
      operationId: getUnreadNotificationsCount
      responses:
        '200':
          description: Количество непрочитанных уведомлений успешно получено
          content:
            application/json:
              schema:
                type: integer
                format: int64
        '401':
          description: Не авторизован
  /notifications/{notificationId}/read:
    put:
      tags:
        - notifications
      summary: Отметка уведомления как прочитанного
      description: Отмечает уведомление как прочитанное
      operationId: markAsRead
      parameters:
        - name: notificationId
          in: path
          description: Идентификатор уведомления
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Уведомление успешно отмечено как прочитанное
        '401':
          description: Не авторизован
        '404':
          description: Уведомление не найдено
  /notifications/read-all:
    put:
      tags:
        - notifications
      summary: Отметка всех уведомлений как прочитанных
      description: Отмечает все уведомления текущего пользователя как прочитанные
      operationId: markAllAsRead
      responses:
        '200':
          description: Все уведомления успешно отмечены как прочитанные
        '401':
          description: Не авторизован
  /notifications/{notificationId}:
    delete:
      tags:
        - notifications
      summary: Удаление уведомления
      description: Удаляет уведомление по его идентификатору
      operationId: deleteNotification
      parameters:
        - name: notificationId
          in: path
          description: Идентификатор уведомления
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Уведомление успешно удалено
        '401':
          description: Не авторизован
        '404':
          description: Уведомление не найдено
...
components:
  schemas:
    ApplicationRequest:
      type: object
      properties:
        eventId:
          type: integer
          format: int64
          description: Идентификатор мероприятия
        profileId:
          type: integer
          format: int64
          description: Идентификатор профиля
          nullable: true
        status:
          type: string
          description: Статус заявки
        submissionDate:
          type: string
          format: date-time
          description: Дата подачи заявки
        message:
          type: string
          description: Сообщение от заявителя
          nullable: true
        isObserver:
          type: boolean
          description: Является ли наблюдателем
        fileApplications:
          type: array
          description: Файлы заявки
          items:
            $ref: '#/components/schemas/FileApplicationRequest'
        verdict:
          type: string
          description: Решение по заявке
          nullable: true
    ApplicationResponse:
      type: object
      properties:
        applicationId:
          type: integer
          format: int64
          description: Идентификатор заявки
        eventId:
          type: integer
          format: int64
          description: Идентификатор мероприятия
        eventName:
          type: string
          description: Название мероприятия
        profileId:
          type: integer
          format: int64
          description: Идентификатор профиля
        status:
          type: string
          description: Статус заявки
        submissionDate:
          type: string
          format: date-time
          description: Дата подачи заявки
        message:
          type: string
          description: Сообщение от заявителя
          nullable: true
        isObserver:
          type: boolean
          description: Является ли наблюдателем
        verdict:
          type: string
          description: Решение по заявке
          nullable: true
        fileApplications:
          type: array
          description: Файлы заявки
          items:
            $ref: '#/components/schemas/FileApplicationResponse'
    ApplicationFileMetadataDTO:
      type: object
      properties:
        type:
          type: string
          description: Тип файла
        isEncrypted:
          type: boolean
          description: Зашифрован ли файл
        docRequiredId:
          type: integer
          format: int64
          description: Идентификатор требуемого документа
...
    FileApplicationRequest:
      type: object
      properties:
        fileName:
          type: string
          description: Название файла
        filePath:
          type: string
          description: Путь к файлу
        uploadDate:
          type: string
          format: date-time
          description: Дата загрузки
        fileType:
          type: string
          description: Тип файла
        isEncryptionEnabled:
          type: boolean
          description: Включено ли шифрование
    FileApplicationResponse:
      type: object
      properties:
        fileId:
          type: integer
          format: int64
          description: Идентификатор файла
        fileName:
          type: string
          description: Название файла
        filePath:
          type: string
          description: Путь к файлу
        uploadDate:
          type: string
          format: date-time
          description: Дата загрузки
        fileType:
          type: string
          description: Тип файла
        docRequired:
          $ref: '#/components/schemas/DocRequiredResponse'
          nullable: true
          description: Требуемый документ
        isEncryptionEnabled:
          type: boolean
          description: Включено ли шифрование
...
    DocRequiredResponse:
      type: object
      properties:
        docRequiredId:
          type: integer
          format: int64
          description: Идентификатор требуемого документа
        name:
          type: string
          description: Название требуемого документа
        description:
          type: string
          description: Описание требуемого документа
          nullable: true
        required:
          type: boolean
          description: Обязателен ли документ
...
    Notification:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор уведомления
        userId:
          type: integer
          format: int64
          description: Идентификатор пользователя
        type:
          type: string
          description: Тип уведомления
          enum:
            - NEW_APPLICATION
            - APPLICATION_STATUS_CHANGED
            - APPLICATION_UPDATED
            - EVENT_UPDATED
            - EVENT_DELETED
            - APPLICATION_DELETED
        title:
          type: string
          description: Заголовок уведомления
        message:
          type: string
          description: Текст уведомления
        isRead:
          type: boolean
          description: Прочитано ли уведомление
        createdAt:
          type: string
          format: date-time
          description: Дата создания
        entityId:
          type: integer
          format: int64
          description: Идентификатор связанной сущности
        entityType:
          type: string
          description: Тип связанной сущности
          enum:
            - EVENT
            - APPLICATION
        status:
          type: string
          description: Статус заявки
          enum:
            - PENDING
            - APPROVED
            - REJECTED
          nullable: true
...
    PageableResponseNotification:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        hasNext:
          type: boolean
          description: Есть ли следующая страница
        totalPages:
          type: integer
          description: Общее количество страниц
        totalElements:
          type: integer
          format: int64
          description: Общее количество элементов
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []